API

API

AstroTime.@timescale — Macro.

@timescale scale

Define a new timescale and the corresponding Epoch type alias.

Example

julia> @timescale GMT ep tai_offset(UTC, ep)

julia> GMT <: TimeScale
true

julia> GMTEpoch
Epoch{GMT,T} where T

source

AstroTime.Epochs.Epoch — Method.

Epoch{S}(ep::Epoch{S}, Δt) where S

Construct a new Epoch with time scale S which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.Epoch — Method.

Epoch{S}(jd1::T, jd2::T=zero(T); origin=:j2000) where {S, T}

Construct an Epoch with time scale S from a Julian date (optionally split into jd1 and jd2).

source

AstroTime.Epochs.TAIEpoch — Type.

Epoch{S}(jd1::T, jd2::T=zero(T); origin=:j2000) where {S, T}

Construct an Epoch with time scale S from a Julian date (optionally split into jd1 and jd2).

Epoch{S}(ep::Epoch{S}, Δt) where S

Construct a new Epoch with time scale S which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.TCBEpoch — Type.

Epoch{S}(jd1::T, jd2::T=zero(T); origin=:j2000) where {S, T}

Construct an Epoch with time scale S from a Julian date (optionally split into jd1 and jd2).

Epoch{S}(ep::Epoch{S}, Δt) where S

Construct a new Epoch with time scale S which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.TCGEpoch — Type.

Epoch{S}(jd1::T, jd2::T=zero(T); origin=:j2000) where {S, T}

Construct an Epoch with time scale S from a Julian date (optionally split into jd1 and jd2).

Epoch{S}(ep::Epoch{S}, Δt) where S

Construct a new Epoch with time scale S which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.TDBEpoch — Type.

Epoch{S}(jd1::T, jd2::T=zero(T); origin=:j2000) where {S, T}

Construct an Epoch with time scale S from a Julian date (optionally split into jd1 and jd2).

Epoch{S}(ep::Epoch{S}, Δt) where S

Construct a new Epoch with time scale S which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.TTEpoch — Type.

Epoch{S}(jd1::T, jd2::T=zero(T); origin=:j2000) where {S, T}

Construct an Epoch with time scale S from a Julian date (optionally split into jd1 and jd2).

Epoch{S}(ep::Epoch{S}, Δt) where S

Construct a new Epoch with time scale S which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.UT1Epoch — Type.

Epoch{S}(jd1::T, jd2::T=zero(T); origin=:j2000) where {S, T}

Construct an Epoch with time scale S from a Julian date (optionally split into jd1 and jd2).

Epoch{S}(ep::Epoch{S}, Δt) where S

Construct a new Epoch with time scale S which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.UTCEpoch — Type.

Epoch{S}(jd1::T, jd2::T=zero(T); origin=:j2000) where {S, T}

Construct an Epoch with time scale S from a Julian date (optionally split into jd1 and jd2).

Epoch{S}(ep::Epoch{S}, Δt) where S

Construct a new Epoch with time scale S which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.now — Method.

now()

Get the current date and time as a UTCEpoch.

source

AstroTime.Epochs.tai_offset — Method.

tai_offset(TDB, ep)

Computes difference TDB-TAI in seconds at the epoch ep.

The accuracy of this routine is approx 40 microseconds in interval 1900-2100 AD. Note that an accurate transformation betweem TDB and TT depends on the trajectory of the observer. For two observers fixed on the earth surface the quantity TDB-TT can differ by as much as about 4 microseconds.

References

https://www.cv.nrao.edu/~rfisher/Ephemerides/times.html#TDB

Issue #26

source