API

API

AstroTime.@timescale — Macro.

@timescale scale

Define a new timescale and the corresponding Epoch type alias.

Example

julia> @timescale GMT ep tai_offset(UTC, ep)

julia> GMT <: TimeScale
true

julia> GMTEpoch
Epoch{GMT,T} where T

source

AstroTime.Epochs.Epoch — Method.

Epoch{S}(ep::Epoch{S}, Δt) where S

Construct a new Epoch with time scale S which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.Epoch — Method.

Epoch{S}(jd1::T, jd2::T=zero(T); origin=:j2000) where {S, T}

Construct an Epoch with time scale S from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> Epoch{UTC}(0.0, 0.5)
2000-01-02T00:00:00.000 UTC

julia> Epoch{UTC}(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 UTC

source

AstroTime.Epochs.TAIEpoch — Method.

TAIEpoch(jd1::T, jd2::T=zero(T); origin=:j2000) where T

Construct a TAIEpoch from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> TAIEpoch(0.0, 0.5)
2000-01-02T00:00:00.000 TAI

julia> TAIEpoch(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 TAI

source

AstroTime.Epochs.TAIEpoch — Method.

TAIEpoch(ep::Epoch{S}, Δt) where S

Construct a TAIEpoch which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.TCBEpoch — Method.

TCBEpoch(jd1::T, jd2::T=zero(T); origin=:j2000) where T

Construct a TCBEpoch from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> TCBEpoch(0.0, 0.5)
2000-01-02T00:00:00.000 TCB

julia> TCBEpoch(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 TCB

source

AstroTime.Epochs.TCBEpoch — Method.

TCBEpoch(ep::Epoch{S}, Δt) where S

Construct a TCBEpoch which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.TCGEpoch — Method.

TCGEpoch(jd1::T, jd2::T=zero(T); origin=:j2000) where T

Construct a TCGEpoch from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> TCGEpoch(0.0, 0.5)
2000-01-02T00:00:00.000 TCG

julia> TCGEpoch(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 TCG

source

AstroTime.Epochs.TCGEpoch — Method.

TCGEpoch(ep::Epoch{S}, Δt) where S

Construct a TCGEpoch which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.TDBEpoch — Method.

TDBEpoch(jd1::T, jd2::T=zero(T); origin=:j2000) where T

Construct a TDBEpoch from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> TDBEpoch(0.0, 0.5)
2000-01-02T00:00:00.000 TDB

julia> TDBEpoch(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 TDB

source

AstroTime.Epochs.TDBEpoch — Method.

TDBEpoch(ep::Epoch{S}, Δt) where S

Construct a TDBEpoch which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.TTEpoch — Method.

TTEpoch(jd1::T, jd2::T=zero(T); origin=:j2000) where T

Construct a TTEpoch from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> TTEpoch(0.0, 0.5)
2000-01-02T00:00:00.000 TT

julia> TTEpoch(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 TT

source

AstroTime.Epochs.TTEpoch — Method.

TTEpoch(ep::Epoch{S}, Δt) where S

Construct a TTEpoch which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.UT1Epoch — Method.

UT1Epoch(jd1::T, jd2::T=zero(T); origin=:j2000) where T

Construct a UT1Epoch from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> UT1Epoch(0.0, 0.5)
2000-01-02T00:00:00.000 UT1

julia> UT1Epoch(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 UT1

source

AstroTime.Epochs.UT1Epoch — Method.

UT1Epoch(ep::Epoch{S}, Δt) where S

Construct a UT1Epoch which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.UTCEpoch — Method.

UTCEpoch(jd1::T, jd2::T=zero(T); origin=:j2000) where T

Construct a UTCEpoch from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> UTCEpoch(0.0, 0.5)
2000-01-02T00:00:00.000 UTC

julia> UTCEpoch(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 UTC

source

AstroTime.Epochs.UTCEpoch — Method.

UTCEpoch(ep::Epoch{S}, Δt) where S

Construct a UTCEpoch which is ep shifted by Δt seconds.

Example

julia> ep = UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(ep, 20.0)
2018-02-06T20:45:20.000 UTC

source

AstroTime.Epochs.now — Method.

now()

Get the current date and time as a UTCEpoch.

source

AstroTime.Epochs.tai_offset — Method.

tai_offset(TDB, ep)

Computes difference TDB-TAI in seconds at the epoch ep.

The accuracy of this routine is approx 40 microseconds in interval 1900-2100 AD. Note that an accurate transformation betweem TDB and TT depends on the trajectory of the observer. For two observers fixed on the earth surface the quantity TDB-TT can differ by as much as about 4 microseconds.

References

https://www.cv.nrao.edu/~rfisher/Ephemerides/times.html#TDB

Issue #26

source