API

API

AstroTime.@timescale — Macro.

@timescale scale

Define a new timescale and the corresponding Epoch type alias.

Example

julia> @timescale GMT ep tai_offset(UTC, ep)

julia> GMT <: TimeScale
true

julia> GMTEpoch
Epoch{GMT,T} where T

source

AstroTime.Epochs.Epoch — Type.

Epoch(str[, format])

Construct an Epoch from a string str. Optionally a format definition can be passed as a DateFormat object or as a string. In addition to the character codes supported by DateFormat the character code D is supported which is parsed as "day of year" (see the example below) and the character code t which is parsed as the time scale.

Note: Please be aware that this constructor requires that the time scale is part of str, e.g. 2018-37T00:00 UTC. Otherwise use the explicit constructor, e.g. Epoch{UTC}.

Example

julia> Epoch("2018-02-06T20:45:00.0 UTC")
2018-02-06T20:45:00.000 UTC

julia> Epoch("2018-37T00:00 UTC", "yyyy-DDDTHH:MM ttt")
2018-02-06T00:00:00.000 UTC

source

AstroTime.Epochs.Epoch — Method.

Epoch{S}(str[, format]) where S

Construct an Epoch with time scale S from a string str. Optionally a format definition can be passed as a DateFormat object or as a string. In addition to the character codes supported by DateFormat the code D can be used which is parsed as "day of year" (see the example below).

Example

julia> Epoch{UTC}("2018-02-06T20:45:00.0")
2018-02-06T20:45:00.000 UTC

julia> Epoch{UTC}("February 6, 2018", "U d, y")
2018-02-06T00:00:00.000 UTC

julia> Epoch{UTC}("2018-37T00:00", "yyyy-DDDTHH:MM")
2018-02-06T00:00:00.000 UTC

source

AstroTime.Epochs.Epoch — Method.

Epoch{S2}(ep::Epoch{S1}) where {S1, S2}

Convert ep, an Epoch with time scale S1, to an Epoch with time scale S2.

Examples

julia> ep = TTEpoch(2000,1,1)
2000-01-01T00:00:00.000 TT

julia> TAIEpoch(ep)
1999-12-31T23:59:27.816 TAI

source

AstroTime.Epochs.Epoch — Method.

Epoch{S}(year, month, day, hour=0, minute=0, second=0.0) where S

Construct an Epoch with time scale S from date and time components.

Example

julia> Epoch{UTC}(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> Epoch{UTC}(2018, 2, 6)
2018-02-06T00:00:00.000 UTC

source

AstroTime.Epochs.Epoch — Method.

Epoch{S}(jd1::T, jd2::T=zero(T); origin=:j2000) where {S, T}

Construct an Epoch with time scale S from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> Epoch{UTC}(0.0, 0.5)
2000-01-02T00:00:00.000 UTC

julia> Epoch{UTC}(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 UTC

source

AstroTime.Epochs.TAIEpoch — Method.

TAIEpoch(str[, format])

Construct a TAIEpoch from a string str. Optionally a format definition can be passed as a DateFormat object or as a string. In addition to the character codes supported by DateFormat the code D is supported which is parsed as "day of year" (see the example below).

Example

julia> TAIEpoch("2018-02-06T20:45:00.0")
2018-02-06T20:45:00.000 TAI

julia> TAIEpoch("February 6, 2018", "U d, y")
2018-02-06T00:00:00.000 TAI

julia> TAIEpoch("2018-37T00:00", "yyyy-DDDTHH:MM")
2018-02-06T00:00:00.000 TAI

source

AstroTime.Epochs.TAIEpoch — Method.

TAIEpoch(year, month, day, hour=0, minute=0, second=0.0)

Construct a TAIEpoch from date and time components.

Example

julia> TAIEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 TAI

julia> TAIEpoch(2018, 2, 6)
2018-02-06T00:00:00.000 TAI

source

AstroTime.Epochs.TAIEpoch — Method.

TAIEpoch(jd1::T, jd2::T=zero(T); origin=:j2000) where T

Construct a TAIEpoch from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> TAIEpoch(0.0, 0.5)
2000-01-02T00:00:00.000 TAI

julia> TAIEpoch(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 TAI

source

AstroTime.Epochs.TCBEpoch — Method.

TCBEpoch(str[, format])

Construct a TCBEpoch from a string str. Optionally a format definition can be passed as a DateFormat object or as a string. In addition to the character codes supported by DateFormat the code D is supported which is parsed as "day of year" (see the example below).

Example

julia> TCBEpoch("2018-02-06T20:45:00.0")
2018-02-06T20:45:00.000 TCB

julia> TCBEpoch("February 6, 2018", "U d, y")
2018-02-06T00:00:00.000 TCB

julia> TCBEpoch("2018-37T00:00", "yyyy-DDDTHH:MM")
2018-02-06T00:00:00.000 TCB

source

AstroTime.Epochs.TCBEpoch — Method.

TCBEpoch(year, month, day, hour=0, minute=0, second=0.0)

Construct a TCBEpoch from date and time components.

Example

julia> TCBEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 TCB

julia> TCBEpoch(2018, 2, 6)
2018-02-06T00:00:00.000 TCB

source

AstroTime.Epochs.TCBEpoch — Method.

TCBEpoch(jd1::T, jd2::T=zero(T); origin=:j2000) where T

Construct a TCBEpoch from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> TCBEpoch(0.0, 0.5)
2000-01-02T00:00:00.000 TCB

julia> TCBEpoch(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 TCB

source

AstroTime.Epochs.TCGEpoch — Method.

TCGEpoch(str[, format])

Construct a TCGEpoch from a string str. Optionally a format definition can be passed as a DateFormat object or as a string. In addition to the character codes supported by DateFormat the code D is supported which is parsed as "day of year" (see the example below).

Example

julia> TCGEpoch("2018-02-06T20:45:00.0")
2018-02-06T20:45:00.000 TCG

julia> TCGEpoch("February 6, 2018", "U d, y")
2018-02-06T00:00:00.000 TCG

julia> TCGEpoch("2018-37T00:00", "yyyy-DDDTHH:MM")
2018-02-06T00:00:00.000 TCG

source

AstroTime.Epochs.TCGEpoch — Method.

TCGEpoch(year, month, day, hour=0, minute=0, second=0.0)

Construct a TCGEpoch from date and time components.

Example

julia> TCGEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 TCG

julia> TCGEpoch(2018, 2, 6)
2018-02-06T00:00:00.000 TCG

source

AstroTime.Epochs.TCGEpoch — Method.

TCGEpoch(jd1::T, jd2::T=zero(T); origin=:j2000) where T

Construct a TCGEpoch from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> TCGEpoch(0.0, 0.5)
2000-01-02T00:00:00.000 TCG

julia> TCGEpoch(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 TCG

source

AstroTime.Epochs.TDBEpoch — Method.

TDBEpoch(str[, format])

Construct a TDBEpoch from a string str. Optionally a format definition can be passed as a DateFormat object or as a string. In addition to the character codes supported by DateFormat the code D is supported which is parsed as "day of year" (see the example below).

Example

julia> TDBEpoch("2018-02-06T20:45:00.0")
2018-02-06T20:45:00.000 TDB

julia> TDBEpoch("February 6, 2018", "U d, y")
2018-02-06T00:00:00.000 TDB

julia> TDBEpoch("2018-37T00:00", "yyyy-DDDTHH:MM")
2018-02-06T00:00:00.000 TDB

source

AstroTime.Epochs.TDBEpoch — Method.

TDBEpoch(year, month, day, hour=0, minute=0, second=0.0)

Construct a TDBEpoch from date and time components.

Example

julia> TDBEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 TDB

julia> TDBEpoch(2018, 2, 6)
2018-02-06T00:00:00.000 TDB

source

AstroTime.Epochs.TDBEpoch — Method.

TDBEpoch(jd1::T, jd2::T=zero(T); origin=:j2000) where T

Construct a TDBEpoch from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> TDBEpoch(0.0, 0.5)
2000-01-02T00:00:00.000 TDB

julia> TDBEpoch(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 TDB

source

AstroTime.Epochs.TTEpoch — Method.

TTEpoch(str[, format])

Construct a TTEpoch from a string str. Optionally a format definition can be passed as a DateFormat object or as a string. In addition to the character codes supported by DateFormat the code D is supported which is parsed as "day of year" (see the example below).

Example

julia> TTEpoch("2018-02-06T20:45:00.0")
2018-02-06T20:45:00.000 TT

julia> TTEpoch("February 6, 2018", "U d, y")
2018-02-06T00:00:00.000 TT

julia> TTEpoch("2018-37T00:00", "yyyy-DDDTHH:MM")
2018-02-06T00:00:00.000 TT

source

AstroTime.Epochs.TTEpoch — Method.

TTEpoch(year, month, day, hour=0, minute=0, second=0.0)

Construct a TTEpoch from date and time components.

Example

julia> TTEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 TT

julia> TTEpoch(2018, 2, 6)
2018-02-06T00:00:00.000 TT

source

AstroTime.Epochs.TTEpoch — Method.

TTEpoch(jd1::T, jd2::T=zero(T); origin=:j2000) where T

Construct a TTEpoch from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> TTEpoch(0.0, 0.5)
2000-01-02T00:00:00.000 TT

julia> TTEpoch(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 TT

source

AstroTime.Epochs.UT1Epoch — Method.

UT1Epoch(str[, format])

Construct a UT1Epoch from a string str. Optionally a format definition can be passed as a DateFormat object or as a string. In addition to the character codes supported by DateFormat the code D is supported which is parsed as "day of year" (see the example below).

Example

julia> UT1Epoch("2018-02-06T20:45:00.0")
2018-02-06T20:45:00.000 UT1

julia> UT1Epoch("February 6, 2018", "U d, y")
2018-02-06T00:00:00.000 UT1

julia> UT1Epoch("2018-37T00:00", "yyyy-DDDTHH:MM")
2018-02-06T00:00:00.000 UT1

source

AstroTime.Epochs.UT1Epoch — Method.

UT1Epoch(year, month, day, hour=0, minute=0, second=0.0)

Construct a UT1Epoch from date and time components.

Example

julia> UT1Epoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UT1

julia> UT1Epoch(2018, 2, 6)
2018-02-06T00:00:00.000 UT1

source

AstroTime.Epochs.UT1Epoch — Method.

UT1Epoch(jd1::T, jd2::T=zero(T); origin=:j2000) where T

Construct a UT1Epoch from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> UT1Epoch(0.0, 0.5)
2000-01-02T00:00:00.000 UT1

julia> UT1Epoch(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 UT1

source

AstroTime.Epochs.UTCEpoch — Method.

UTCEpoch(str[, format])

Construct a UTCEpoch from a string str. Optionally a format definition can be passed as a DateFormat object or as a string. In addition to the character codes supported by DateFormat the code D is supported which is parsed as "day of year" (see the example below).

Example

julia> UTCEpoch("2018-02-06T20:45:00.0")
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch("February 6, 2018", "U d, y")
2018-02-06T00:00:00.000 UTC

julia> UTCEpoch("2018-37T00:00", "yyyy-DDDTHH:MM")
2018-02-06T00:00:00.000 UTC

source

AstroTime.Epochs.UTCEpoch — Method.

UTCEpoch(year, month, day, hour=0, minute=0, second=0.0)

Construct a UTCEpoch from date and time components.

Example

julia> UTCEpoch(2018, 2, 6, 20, 45, 0.0)
2018-02-06T20:45:00.000 UTC

julia> UTCEpoch(2018, 2, 6)
2018-02-06T00:00:00.000 UTC

source

AstroTime.Epochs.UTCEpoch — Method.

UTCEpoch(jd1::T, jd2::T=zero(T); origin=:j2000) where T

Construct a UTCEpoch from a Julian date (optionally split into jd1 and jd2). origin determines the variant of Julian date that is used. Possible values are:

:j2000: J2000 Julian date, starts at 2000-01-01T12:00

:julian: Julian date, starts at -4712-01-01T12:00

:mjd: J2000 Julian date, starts at 1858-11-17T00:00

Examples

julia> UTCEpoch(0.0, 0.5)
2000-01-02T00:00:00.000 UTC

julia> UTCEpoch(2.451545e6, origin=:julian)
2000-01-01T12:00:00.000 UTC

source

AstroTime.Epochs.now — Method.

now()

Get the current date and time as a UTCEpoch.

source

AstroTime.Epochs.tai_offset — Method.

tai_offset(TDB, ep)

Computes difference TDB-TAI in seconds at the epoch ep.

The accuracy of this routine is approx 40 microseconds in interval 1900-2100 AD. Note that an accurate transformation betweem TDB and TT depends on the trajectory of the observer. For two observers fixed on the earth surface the quantity TDB-TT can differ by as much as about 4 microseconds.

References

https://www.cv.nrao.edu/~rfisher/Ephemerides/times.html#TDB

Issue #26

source