API

API

AstronomicalTime.@timescale — Macro.

@timescale scale

Define a new timescale and the corresponding Epoch type alias.

Example

julia> @timescale Custom

julia> Custom <: TimeScale
true
julia> CustomEpoch == Epoch{Custom, T} where T <: Number
true

source

AstronomicalTime.Epochs.Epoch — Method.

Epoch{T}(year, month, day,
    hour=0, minute=0, seconds=0, milliseconds=0) where T<:TimeScale

Construct an Epoch with timescale T at the given date and time.

Example

julia> Epoch{TT}(2017, 3, 14, 7, 18, 20, 325)
2017-03-14T07:18:20.325 TT

source

AstronomicalTime.Epochs.Epoch — Method.

Epoch{T}(timestamp::AbstractString,
    fmt::DateFormat=dateformat"yyyy-mm-ddTHH:MM:SS.sss") where T<:TimeScale

Construct an Epoch with timescale T from a timestamp. Optionally a DateFormat object can be passed which improves performance if many date strings need to be parsed and the format is known in advance.

Example

julia> Epoch{TT}("2017-03-14T07:18:20.325")
2017-03-14T07:18:20.325 TT

source

AstronomicalTime.Epochs.Epoch — Method.

Epoch{T}(ep::Epoch{S}) where {T<:TimeScale, S<:TimeScale}

Convert an Epoch with timescale S to an Epoch with timescale T.

Example

julia> Epoch{TT}(Epoch{TAI}(2000, 1, 1))
2000-01-01T00:00:32.184 TT

source

AstronomicalTime.Epochs.Epoch — Method.

Epoch{T}(dt::DateTime) where T<:TimeScale

Convert a DateTime object to an Epoch with timescale T.

Example

julia> Epoch{TT}(DateTime(2017, 3, 14, 7, 18, 20, 325))
2017-03-14T07:18:20.325 TT

source