Time Scales

AstroTime.update — Function

AstroTime.update()

Download up-to-date IERS tables from the internet to enable transformations to and from the UT1 time scale.

AstroTime.@timescale — Macro

@timescale scale [parent[, oneway]]

Define a new time scale and the corresponding Epoch type alias.

Arguments

scale: The name of the time scale

parent: The "parent" time scale to which it should be linked (optional)

oneway: If true, only the transformation from parent to scale is registered (optional, default: false)

Example

julia> @timescale GMT TAI

julia> GMT isa TimeScale
true

julia> GMTEpoch
Epoch{GMTScale}

julia> find_path(TT, GMT)
3-element Vector{TimeScale}:
 TT
 TAI
 GMT

AstroTime.TimeScales.TAI — Constant

TAI

The singleton instance of the InternationalAtomicTime type representing the International Atomic Time (TAI) time scale.

References

Wikipedia

AstroTime.TimeScales.TCB — Constant

TCB

The singleton instance of the BarycentricCoordinateTime type representing the Barycentric Coordinate Time (TCB) time scale.

References

Wikipedia

AstroTime.TimeScales.TCG — Constant

TCG

The singleton instance of the GeocentricCoordinateTime type representing the Geocentric Coordinate Time (TCG) time scale.

References

Wikipedia

AstroTime.TimeScales.TDB — Constant

TDB

The singleton instance of the BarycentricDynamicalTime type representing the Barycentric Dynamical Time (TDB) time scale.

References

Wikipedia

AstroTime.TimeScales.TT — Constant

TT

The singleton instance of the TerrestrialTime type representing the Terrestrial Time (TT) time scale.

References

Wikipedia

AstroTime.TimeScales.UT1 — Constant

UT1

The singleton instance of the UniversalTime type representing the Universal Time (UT1) time scale.

References

Wikipedia

AstroTime.TimeScales.BarycentricCoordinateTime — Type

BarycentricCoordinateTime

A type representing the Barycentric Coordinate Time (TCB) time scale.

References

Wikipedia

AstroTime.TimeScales.BarycentricDynamicalTime — Type

BarycentricDynamicalTime

A type representing the Barycentric Dynamical Time (TDB) time scale.

References

Wikipedia

AstroTime.TimeScales.GeocentricCoordinateTime — Type

GeocentricCoordinateTime

A type representing the Geocentric Coordinate Time (TCG) time scale.

References

Wikipedia

AstroTime.TimeScales.InternationalAtomicTime — Type

InternationalAtomicTime

A type representing the International Atomic Time (TAI) time scale.

References

Wikipedia

AstroTime.TimeScales.TerrestrialTime — Type

TerrestrialTime

A type representing the Terrestrial Time (TT) time scale.

References

Wikipedia

AstroTime.TimeScales.TimeScale — Type

All timescales are subtypes of the abstract type TimeScale. The following timescales are defined:

UT1 — Universal Time

TAI — International Atomic Time

TT — Terrestrial Time

TCG — Geocentric Coordinate Time

TCB — Barycentric Coordinate Time

TDB — Barycentric Dynamical Time

AstroTime.TimeScales.UniversalTime — Type

UniversalTime

A type representing the Universal Time (UT1) time scale.

References

Wikipedia